% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_n_unique_vals.R
\name{find_n_unique_vals}
\alias{find_n_unique_vals}
\title{Find n unique values within a data frame.}
\usage{
find_n_unique_vals(.data, n_unique, return_tibble = TRUE)
}
\arguments{
\item{.data}{a table of data}

\item{n_unique}{a numeric vector of length 1 or greater.}

\item{return_tibble}{boolean, if TRUE, a data frame is returned; if false, a list is returned.}
}
\value{
a data frame with column names, unique values, and number of unique values
}
\description{
Find n unique values within a data frame.
}
\examples{
find_n_unique_vals(mtcars, 2:3)
}
