% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unique_vals.R
\name{find_unique_vals}
\alias{find_unique_vals}
\title{Find unique values in a data frame}
\usage{
find_unique_vals(.data, pattern = NULL, return_tibble = TRUE)
}
\arguments{
\item{.data}{a table of data}

\item{pattern}{Pattern to look for}

\item{return_tibble}{boolean, if TRUE, a data frame is returned; if false, a list is returned.}
}
\value{
a data frame with column names, unique values, and number of unique values
}
\description{
Find unique values in a data frame
}
\examples{
\dontrun{
library(gapminder)
find_unique_vals(gapminder, "Kuwait")
}
}
