% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_vals.R
\name{get_unique_vals}
\alias{get_unique_vals}
\title{Get unique values by column}
\usage{
get_unique_vals(.data, ...)
}
\arguments{
\item{.data}{a table of data.}

\item{...}{One or more unquoted expressions separated by commas. Can pass helper functions, e.g. starts_with(), contains(), etc.}
}
\value{
a data frame with column names, unique values, and number of unique values
}
\description{
Get unique values by column
}
